Improve behavior of build scripts
This commit takes care of the possible increase of pdflatex
iterations introduced by the previous commit ("qqz: Improve
accuracy of cross-links") that would cause 'make' to terminate
prematurely. The same issue seems to have been observed since
commit 33b93f8258f5 ("qqz: Cross-link questions and answers").
However, no one has attempted to fix it because a final PDF can
be obtained by just repeating 'make' again.

To fix the issue, this commit refactors 'runlinux.sh' into 3 parts.

1) bibtex related part is moved to 'Makefile' as rules for targets
such as 'perfbook.bbl'. It also adds necessary rules which support
the '.bbl' rules.

2) 'runfirstlatex.sh' is separated so that it can be invoked in
rules for '.aux' targets which support '.bbl' rules.

3) 'runlatex.sh' is intensively modified. The first while loop
watches the 'LaTeX Warning: There were undefined references'
message and the second one watches the 'LaTeX Warning: Label(s)
may have changed' message.
The loops watch the variation in 'LaTeX Warning:" lines in
$basename.log and when they find no change, they will give up.
Note that in early rounds of the loop, there may be cases when
there is no change in 'LaTeX Warning' messages while there
is still room for improvement. To accommodate this, giving-up
will fire only after a certain number of iterations for each loop.
Since the minimum iteration numbers are chosen after trial and
error, they might need to be changed in the future.

For each round of the loop, a message representing which warning
is being watched is displayed.
When the giving-up logic kicks in, grep of the remaining warnings
is also displayed.

'$basename-first.log', '$basename-warning.log' and
'$basename-warning-prev.log' files are used to track things.
They are removed when the build succeeds.

This commit also updates the copyright notices in the scripts.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>