Remove *.bbl from repository
As rules for '*.bbl' targets are added to 'Makefile' in the
previous commit ("Improve behavior of build scripts"), they are
no longer necessary in the repository. This commit removes them.
It also updates .gitignore and the 'make clean' rule accordingly.
It also adds 'perfbook-hb.out' to the 'make clean' rule.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>